FacePrivacy/PRO-Face D/requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.4.1
async-timeout==4.0.3
attrs==23.2.0
blinker==1.8.2
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.1.1
cycler==0.12.1
dlib @ file:///home/conda/feedstock_root/build_artifacts/dlib-split_1712214213590/work
easydict==1.13
face-recognition==1.3.0
face-recognition-models==0.3.0
facenet-pytorch==2.6.0
filelock==3.16.1
Flask==3.0.3
Flask-Cors==5.0.0
Flask-SQLAlchemy==3.1.1
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.3.1
ftfy==6.2.0
future==1.0.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.29.0
google-auth-oauthlib==1.0.0
greenlet==3.1.1
grpcio==1.63.0
idna==3.7
importlib_metadata==7.1.0
importlib_resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lmdb==1.2.1
lpips==0.1.4
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
narwhals==1.13.3
networkx==3.1
numpy @ file:///home/conda/feedstock_root/build_artifacts/numpy_1687808301083/work
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
opencv-python==4.11.0.86
packaging==24.0
pandas==2.0.3
pillow==10.2.0
pkgutil_resolve_name==1.3.10
protobuf==5.26.1
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydeck==0.9.1
pyDeprecate==0.3.1
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytorch-fid==0.2.0
pytorch-lightning==1.4.5
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
regex==2024.5.10
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.9.4
rpds-py==0.20.1
rsa==4.9
scipy==1.5.4
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.38
streamlit==1.40.0
sympy==1.13.3
tenacity==9.0.0
tensorboard==2.14.0
tensorboard-data-server==0.7.2
toml==0.10.2
torch==2.2.2
torchaudio @ file:///home/zzs/torchaudio-0.9.1-cp38-cp38-linux_x86_64.whl#sha256=b5c58c4765077bb1fc8c1f5df6d11a6406f658185889ea2ee1d996f89061f465
torchmetrics==0.5.0
torchvision==0.17.2
tornado==6.4.1
tqdm==4.66.4
triton==2.2.0
typing_extensions==4.11.0
tzdata==2024.2
urllib3==2.2.1
watchdog==4.0.2
wcwidth==0.2.13
Werkzeug==3.0.3
yarl==1.9.4
zipp==3.18.1
